feat: remove node 4.x support
With the introduction of `@ngtools/json-schema` a triple dot array spread operator (`...[1, 2, 3]`) comes into the code base. This syntax is not supported on node 4.x

As this is the future, we consequently drop support for old node versions. Please update your node installations if you wish to continue using ng-packagr!