test: improve quality and coverage in integration testing
Verifies the public API surface of some FESM'5 and UMD bundles.  Verifies Ahead-of-Time (AOT) metadata in version 4 forAngular v5. Enables to run tests for an individual sample.

Let not-so-good tests slip in as I was accepting features and pull requests to get the project forward and become broader reach. Before refactoring, need to improve current working samples by adding tests for them.