feat: dynamic rollup configuration
BREAKING CHANGE:
`externals` has been removed in favor of `umdModuleIds` which is now only used to name the external libraries.

By default all dependencies are now treated as externals and thus are not embedded in the final bundle.

However, If you want to embed one or more of them into the final bundle you can do so by adding the library in the `embeeded` section like so;

```json
{
  "$schema": "../../../src/ng-package.schema.json",
  "lib": {
    "embedded": [
      "lodash",
      "date-fns"
    ]
  }
}
```

While `rollup` will do it's best to guess the name of an external library some modules are hard to guess and it's recommended to provide the name of the external dependency. This can be done by using `umdModuleIds` in the config section like so;

```json
{
  "$schema": "../../../src/ng-package.schema.json",
  "lib": {
    "umdModuleIds": {
      "lodash" : "_",
      "date-fns" : "DateFns",
    }
  }
}
```

Closes: #312