feat: also create package .tgz
This PR causes `ng-packagr` to also create a `.tgz` file for the library being packaged.

`npm install path/to/directory` creates symlinks since `npm@5.0.0` (see the feature summary in https://github.com/npm/npm/releases/tag/v5.0.0).

The symlink will cause library dependencies to be resolved to in the library project instead of the consuming application.

This is problematic with peer dependencies in general and especially Angular libraries, since the application and the library are using different instances of `@angular/core`. This in turn causes hard to debug errors like the infamous `is not an NgModule` errors.

The correct way to install built packages is to install from the .tgz file. This will not create a symlink and instead install the package as if it was from the npm registry, resolving all dependencies in the consumed project.