fix: create an array large enough to hold all buckets
This will fix an error that appears only when using multiple secondary entry-points depending on each other.
Updated the examples to cause the error and fixed it with a proper initialization of the involved array holding our buckets.

That's the error it will fix:
```
Cannot read property 'push' of undefined
TypeError: Cannot read property 'push' of undefined
    at nodeDepths.forEach (/Users/geokal/job/opensource/ng-packagr/src/lib/brocc/depth.ts:107:22)
```