feat: add support for incremental builds
This PR, bring incremental builds, in `CLI` you can enable the watch, by providing the `-w` or `--watch` arguments. While for the `API`, you can use the `withOptions` method example.

CLI
```
ng-packagr --project ng-packagr.js --watch
```

API
```
ngPackage
  .ngPackagr()
  .forProject('ng-package.js')
  .withOptions({watch: true})
  .build()
```

This requires `enableResourceInlining` to be enabled and also this is a sort of MVP. There are more optimization that need to be done.

Such as;
- Cache Compiler Hosts for later usage
- Cache Program for later usage
- Process Stylesheet at an earlier stage
- Improve depencency analyses by a bit more
- Provide a callback on every re-compilation

Closes: #828, Closes: #743 and Closes: #635