feat: update `lib` compiler option to `es2018`
BREAKING CHANGE:

The following `peerDependencies` are now required
- `tslib: ^1.9.0`
- `typescript: ^2.7.0`

`languageLevel` option has been dropped from the configuration, in favor of the standard tsconfig option `lib`. The default is now `es2018`. In case this needs to be changed. This can to be done by overriding the tsconfig. More on how to override the tsconfig can be found here: https://github.com/dherges/ng-packagr/blob/master/docs/override-tsconfig.md