feat: show error when trying to publish ivy packages
Publishing of Ivy packages should not be allowed, as there Ivy packages are not backward compatible with VE applications.

With this change we use the npm/yarn prepublishOnly hook to display and error and abort the process with a non zero error code when a user tries to publish an Ivy version of the package.

More info: https://docs.npmjs.com/misc/scripts