fix: do not run ngcc when node_modules does not exist
Prior to this change, ng-packagr would fail with an error when ngcc could not
be run due to a missing `node_modules` directory, as is the case with Yarn
Plug'n'Play for example. The workaround was to create an empty `node_modules`
directory, but this is inconvenient and shouldn't make any difference.

This commit removes the error and simply skips ngcc processing.