perf: reduce TypeScript JSDoc parsing
TypeScript 5.3 provides a new programmatically accessible option on the compiler host object to control the amount of JSDoc parsing that the TypeScript parser will perform. The tsc command line tool now uses the ParseForTypeErrors value which only parses JSDoc comments that may affect type checking and is considered a good default for tools such as the Angular CLI. ng-packagr will now attempt to use the ParseForTypeErrors value as well when available. Projects will need to use TypeScript 5.3+ for this option to be available. No behavior changes will occur on TypeScript 5.2 projects. This should not only provide a small improvement to build times but also a reduction in overall memory usage.