refactor: disable Piscina timing information recording
The latest version of the worker pool package used by the build system (Piscina) now has an option to disable performance timing information. Since the build system does not currently use that information, disabling it avoids the extra overhead of histogram creation and usage. This information collection could potentially be enabled conditionally in the future if needed.