feat: add support for node module resolution in style @use and @import
Currently the import resolver does not work in pnpm repositories, since for example if you install
@angular/material and import something from it, it will not find the @material/typography package,
since it is not symlinked to the root node_modules directory. Instead it can be found relative to the
@angular/material location, for example node_modules/.pnpm/@angular+material@version/node_modules/@material/typography.
This commit add a FileImporter which follows the node module resolution, an therefore find the imports correctly.